requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
psycopg2-binary
Werkzeug
scikit-learn
numpy
pandas
python-dotenv